LIA/LIA1/requirements.txt
Flask==2.2.2
Pillow